+++ udata/rdf.py
@@ -22,6 +22,8 @@ SPDX = Namespace('http://spdx.org/rdf/terms#')
 VCARD = Namespace('http://www.w3.org/2006/vcard/ns#')
 FREQ = Namespace('http://purl.org/cld/freq/')
 EUFREQ = Namespace('http://publications.europa.eu/resource/authority/frequency/')  # noqa: E501
+EUFORMAT = Namespace('http://publications.europa.eu/resource/authority/file-type/')
+IANAFORMAT = Namespace('https://www.iana.org/assignments/media-types/')
 DCT = DCTERMS  # More common usage
 
 namespace_manager = NamespaceManager(Graph())